move store slices into `store/slices` directory
