+++ frontend/src/common/test-utils.tsx
@@ -15,8 +15,8 @@ import { Provider } from "react-redux";
 
 import { localBackendURL } from "@/common/test-constants";
 import { Faces } from "@/common/types";
-import { setURL } from "@/features/backend/backendSlice";
 import { LayoutWithoutReduxProvider } from "@/features/ui/Layout";
+import { setURL } from "@/store/slices/backendSlice";
 import { RootState, setupStore } from "@/store/store";
 
 //# region redux test setup
