+++ frontend/src/features/import/ImportText.tsx
@@ -28,9 +28,9 @@ import {
 import { useAppDispatch, useAppSelector } from "@/common/types";
 import { toTitleCase } from "@/common/utils";
 import { RightPaddedIcon } from "@/components/icon";
-import { addMembers, selectProjectSize } from "@/features/project/projectSlice";
-import { selectFuzzySearch } from "@/features/searchSettings/SearchSettingsSlice";
 import { useGetDFCPairsQuery, useGetSampleCardsQuery } from "@/store/api";
+import { addMembers, selectProjectSize } from "@/store/slices/projectSlice";
+import { selectFuzzySearch } from "@/store/slices/SearchSettingsSlice";
 
 export function ImportText() {
   //# region queries and hooks
