+++ frontend/src/pages/index.tsx
@@ -11,13 +11,13 @@ import {
   MakePlayingCardsURL,
   ProjectName,
 } from "@/common/constants";
-import {
-  useBackendConfigured,
-  useProjectName,
-} from "@/features/backend/backendSlice";
 import { DynamicLogo } from "@/features/ui/DynamicLogo";
 import Footer from "@/features/ui/Footer";
 import { ProjectContainer } from "@/features/ui/Layout";
+import {
+  useBackendConfigured,
+  useProjectName,
+} from "@/store/slices/backendSlice";
 
 function JumpIntoEditorButton() {
   const backendConfigured = useBackendConfigured();
