+++ frontend/src/store/slices/SearchSettingsSlice.ts
