+++ frontend/src/store/slices/viewSettingsSlice.ts
